Refactor code out of AppInterface
- Move sidebar stuff into SideBar.js
- Move TripViewContent stuff into that file
- Make Breakfast uppercase
