+++ client/src/components/Timeline/index.js
@@ -1,1 +1,2 @@
 export { TeaCup } from './TeaCup'
+export { TripViewContent } from './TripViewContent'
