+++ client/src/components/sideBar/SideBar.js
@@ -4,7 +4,12 @@ import AddIcon from '@mui/icons-material/Add'
 
 import { Button, Toggle } from '../common'
 import { AppView } from '../../constants/enums'
-import { setActiveTripId, setAppView } from '../../redux/reducers/viewSlice'
+import {
+  closeSidebar,
+  setActiveTripId,
+  setAppView,
+  toggleSidebar,
+} from '../../redux/reducers/viewSlice'
 import { Logout } from '../user'
 import {
   toggleCompactView,
@@ -15,6 +20,7 @@ export function SideBar() {
   const dispatch = useDispatch()
   const appView = useSelector((state) => state.view.appView)
   const activeTripId = useSelector((state) => state.view.activeTripId)
+  const isSidebarOpen = useSelector((state) => state.view.isSidebarOpen)
   const isCompactView = useSelector((state) => state.preferences.compactView)
   const isVerticalTimelines = useSelector(
     (state) => state.preferences.verticalTimelines,
@@ -68,40 +74,101 @@ export function SideBar() {
   const preferencesModalBtn = useMemo(() => {
     // TODO refactor this to return a button that opens a modal.
     //    Please follow any conventions that Meng uses when styling / storing modal content
+    const compactViewToggle = (
+      <Toggle
+        className='mb-2'
+        label='Compact View'
+        onClick={() => {
+          dispatch(toggleCompactView())
+        }}
+        active={isCompactView}
+      />
+    )
+    const verticalTimelinesToggle = (
+      <Toggle
+        label='Vertical Timelines'
+        onClick={() => {
+          dispatch(toggleVerticalTimelines())
+        }}
+        active={isVerticalTimelines}
+      />
+    )
+
     return (
       <div className='mb-2 w-full rounded-md bg-slate-100/40 p-2'>
         <h3 className='mb-1 text-lg font-semibold'>Preferences</h3>
-        <Toggle
-          className='mb-2'
-          label='Compact View'
-          onClick={() => {
-            dispatch(toggleCompactView())
-          }}
-          active={isCompactView}
-        />
-        <Toggle
-          label='Vertical Timelines'
-          onClick={() => {
-            dispatch(toggleVerticalTimelines())
-          }}
-          active={isVerticalTimelines}
-        />
+        {compactViewToggle}
+        {verticalTimelinesToggle}
       </div>
     )
   }, [isVerticalTimelines, isCompactView])
 
-  return (
-    <div className="flex h-full w-full flex-col justify-between overflow-hidden bg-slate-300 bg-[url('../public/little-prince.jpg')] bg-cover bg-center p-2 bg-blend-soft-light">
+  const renderSidebarMainPortion = useMemo(() => {
+    return (
       <div className='overflow-x-auto'>
         {newTripButton}
         <div className='my-2 grid w-full grid-cols-1 gap-2 border-y-2 border-slate-300 py-2'>
           {tripEntries}
         </div>
       </div>
+    )
+  }, [newTripButton, tripEntries])
+
+  const renderSidebarBottomPortion = useMemo(() => {
+    return (
       <div className='box-border w-full shrink-0 py-2'>
         {preferencesModalBtn}
         <Logout />
       </div>
-    </div>
+    )
+  }, [preferencesModalBtn])
+
+  const renderSidebar = useMemo(() => {
+    return (
+      <div
+        className={`fixed left-0 top-0 z-50 flex h-full w-1/5 flex-row overflow-hidden transition-all
+                       ${isSidebarOpen ? 'left-0' : 'left-[-20vw]'}`}
+      >
+        <div className="flex h-full w-full flex-col justify-between overflow-hidden bg-slate-300 bg-[url('../public/little-prince.jpg')] bg-cover bg-center p-2 bg-blend-soft-light">
+          {renderSidebarMainPortion}
+          {renderSidebarBottomPortion}
+        </div>
+      </div>
+    )
+  }, [isSidebarOpen, renderSidebarMainPortion, renderSidebarBottomPortion])
+
+  const renderSidebarToggleButton = useMemo(() => {
+    return (
+      <span
+        className={`fixed top-0 z-50 flex h-full items-center rounded-none transition-all ${
+          isSidebarOpen ? 'left-[20vw]' : 'left-0'
+        }`}
+      >
+        <Button
+          onClick={() => dispatch(toggleSidebar())}
+          className='h-full w-10 rounded-none bg-slate-300 text-6xl hover:bg-slate-400'
+        >
+          {isSidebarOpen ? '‹' : '›'}
+        </Button>
+      </span>
+    )
+  }, [isSidebarOpen, dispatch])
+
+  const renderSidebarShading = useMemo(() => {
+    if (!isSidebarOpen) return <></>
+    return (
+      <div
+        className={`fixed left-0 top-0 z-[45] h-full w-full bg-slate-800 opacity-30`}
+        onClick={() => dispatch(closeSidebar())}
+      />
+    )
+  }, [dispatch, isSidebarOpen])
+
+  return (
+    <>
+      {renderSidebar}
+      {renderSidebarToggleButton}
+      {renderSidebarShading}
+    </>
   )
 }
